Resolve TODOs
