+++ packages/signed-api/src/handlers.ts
@@ -1,98 +1,23 @@
-import {
-  deriveOevTemplateId,
-  type SignedApiBatchPayloadV1,
-  signedApiBatchPayloadV1Schema,
-  type SignedApiBatchPayloadV2,
-  signedApiBatchPayloadV2Schema,
-} from '@api3/airnode-feed';
-import { deriveBeaconId, type Hex } from '@api3/commons';
+import { signedApiBatchPayloadV1Schema, signedApiBatchPayloadV2Schema } from '@api3/airnode-feed';
 import { go, goSync } from '@api3/promise-utils';
 import { isEmpty, omit, pick } from 'lodash';
 
 import { getConfig } from './config/config';
 import { loadEnv } from './env';
 import { createResponseHeaders } from './headers';
-import { get, getAll, getAllAirnodeAddresses, getCache, prune, putAll, setCache } from './in-memory-cache';
+import { get, getAll, getAllAirnodeAddresses, prune, putAll } from './in-memory-cache';
 import { logger } from './logger';
-import { evmAddressSchema, type InternalSignedData, type Endpoint } from './schema';
+import { type Endpoint, evmAddressSchema, type InternalSignedData } from './schema';
 import { getVerifier } from './signed-data-verifier-pool';
+import { transformAirnodeFeedPayload } from './transform-payload';
 import type {
   ApiResponse,
-  PostSignedDataResponseSchema,
+  GetListAirnodesResponseSchema,
   GetSignedDataResponseSchema,
   GetUnsignedDataResponseSchema,
-  GetListAirnodesResponseSchema,
+  PostSignedDataResponseSchema,
 } from './types';
 import { extractBearerToken, generateErrorResponse, isBatchUnique } from './utils';
-
-// TODO: Move from handlers
-export const getOevTemplateId = (templateId: string) => {
-  const cache = getCache();
-  if (templateId in cache.templateIdToOevTemplateId) {
-    return cache.templateIdToOevTemplateId[templateId]!;
-  }
-
-  const oevTemplateId = deriveOevTemplateId(templateId);
-  setCache({
-    ...cache,
-    templateIdToOevTemplateId: { ...cache.templateIdToOevTemplateId, [templateId]: oevTemplateId },
-  });
-  return oevTemplateId;
-};
-
-// TODO: Move from handlers
-export const getBeaconId = (airnode: string, templateId: string) => {
-  const cache = getCache();
-  if (cache.airnodeToTemplateIdToBeaconId[airnode]?.[templateId]) {
-    return cache.airnodeToTemplateIdToBeaconId[airnode][templateId];
-  }
-
-  const beaconId = deriveBeaconId(airnode as Hex, templateId as Hex);
-  setCache({
-    ...cache,
-    airnodeToTemplateIdToBeaconId: {
-      ...cache.airnodeToTemplateIdToBeaconId,
-      [airnode]: { ...cache.airnodeToTemplateIdToBeaconId[airnode], [templateId]: beaconId },
-    },
-  });
-  return beaconId;
-};
-
-// TODO: Move from handlers
-export const transformAirnodeFeedPayload = (
-  payload: SignedApiBatchPayloadV1 | SignedApiBatchPayloadV2
-): InternalSignedData[] => {
-  if ('airnode' in payload) {
-    const { airnode } = payload;
-    return payload.signedData.flatMap((data) => {
-      const { templateId, encodedValue, oevSignature, signature, timestamp } = data;
-      const oevTemplateId = getOevTemplateId(templateId);
-      return [
-        {
-          airnode,
-          beaconId: getBeaconId(airnode, templateId),
-          encodedValue,
-          signature,
-          templateId,
-          timestamp,
-          isOevBeacon: false,
-        },
-        {
-          airnode,
-          beaconId: getBeaconId(airnode, oevTemplateId),
-          encodedValue,
-          signature: oevSignature,
-          templateId: oevTemplateId,
-          timestamp,
-          isOevBeacon: true,
-        },
-      ];
-    });
-  }
-
-  // Airnode feed v1 does not send the data for OEV beacons.
-  return payload.map((data) => ({ ...data, isOevBeacon: false }));
-};
 
 const env = loadEnv();
 // Accepts a batch of signed data that is first validated for consistency and data integrity errors. If there is any
